同步汉化07-31

+++ code/modules/reagents/chemistry/reagents/other_reagents.dm
@@ -316,6 +316,7 @@
 	randomized_spawns = REAGENT_SPAWN_ALL_RANDOM_SPAWNS
 	default_container = /obj/item/reagent_containers/cup/glass/bottle/holywater
 	metabolized_traits = list(TRAIT_HOLY)
+	COOLDOWN_DECLARE(spell_clear_cd)
 
 /datum/glass_style/drinking_glass/holywater
 	required_drink_type = /datum/reagent/water/holywater
@@ -351,14 +352,26 @@
 	affected_mob.adjust_jitter_up_to(2 SECONDS * metabolization_ratio * seconds_per_tick, 20 SECONDS)
 	var/need_mob_update = FALSE
 
-	if(IS_CULTIST(affected_mob))
-		for(var/datum/action/innate/cult/blood_magic/BM in affected_mob.actions)
+	if(COOLDOWN_FINISHED(src, spell_clear_cd))
+		if(IS_CULTIST(affected_mob))
+			for(var/datum/action/innate/cult/blood_magic/BM in affected_mob.actions)
+				var/removed_any = FALSE
+				for(var/datum/action/innate/cult/blood_spell/BS in BM.spells)
+					removed_any = TRUE
+					qdel(BS)
+				if(removed_any)
+					to_chat(affected_mob, span_cult_large(LANG("datum.9e5484f2", null)))
+					COOLDOWN_START(src, spell_clear_cd, 3 SECONDS)
+
+		if(IS_HERETIC(affected_mob))
+			var/datum/antagonist/heretic/heretic_datum = GET_HERETIC(affected_mob)
 			var/removed_any = FALSE
-			for(var/datum/action/innate/cult/blood_spell/BS in BM.spells)
-				removed_any = TRUE
-				qdel(BS)
+			for(var/datum/heretic_knowledge/spell/spell in heretic_datum.get_researched_knowledge())
+				if(spell.remove_charges(ceil(spell.max_charges * spell.holywater_drain_amount)))
+					removed_any = TRUE
 			if(removed_any)
-				to_chat(affected_mob, span_cult_large("Your blood rites falter as holy water scours your body!"))
+				to_chat(affected_mob, span_mansus(LANG("datum.e0377bfb", null)))
+				COOLDOWN_START(src, spell_clear_cd, 3 SECONDS)
 
 	if(data["deciseconds_metabolized"] >= (25 SECONDS)) // 10 units
 		affected_mob.adjust_stutter_up_to(2 SECONDS * metabolization_ratio * seconds_per_tick, 20 SECONDS)
@@ -3448,7 +3461,7 @@
 	chemical_flags = REAGENT_CAN_BE_SYNTHESIZED
 	randomized_spawns = REAGENT_SPAWN_ALL_RANDOM_SPAWNS
 	overdose_threshold = 50 // GLOW GLOW GLOW
-	metabolized_traits = list(TRAIT_MINOR_NIGHT_VISION)
+	metabolized_traits = list(TRAIT_NIGHT_VISION)
 	self_consuming = TRUE
 	/// Fake flashlight we're using to make owner's eyes glow
 	var/obj/item/flashlight/eyelight/glow/glowing
@@ -3528,7 +3541,7 @@
 	name = "Red Luminiscent Fluid"
 	color = COLOR_SOFT_RED
 	// The glow *is* unnatural, so...
-	metabolized_traits = list(TRAIT_MINOR_NIGHT_VISION, TRAIT_UNNATURAL_RED_GLOWY_EYES)
+	metabolized_traits = list(TRAIT_NIGHT_VISION, TRAIT_UNNATURAL_RED_GLOWY_EYES)
 
 /datum/reagent/luminescent_fluid/red/overdose_start(mob/living/affected_mob, metabolization_ratio)
 	. = ..()
